Refactor remaining questions to use BaseQuestionApp
[#172821069]
