+++ src/fill-in-the-blank/components/runtime.scss
@@ -1,9 +1,5 @@
 @import "../../shared/styles/helpers";
 
-.runtime {
-  @include lara-styles;
-}
-
 .correctAnswer {
   color: $correctColor;
 }
