+++ src/multiple-choice/components/app.test.tsx
@@ -1,35 +1,41 @@
 import React from "react";
-import { mount, shallow } from "enzyme";
-import { App } from "./app";
-import { Runtime } from "./runtime";
-import { Authoring } from "./authoring";
+import { baseAuthoringProps, IChoice } from "./app";
 
-let mode: any;
-jest.mock("../../shared/hooks/use-lara-interactive-api", () => ({
-    useLARAInteractiveAPI: () => ({ mode, authoredState: {} })
-  })
-);
+describe("preprocessFormData helper", () => {
+  it("generates unique IDs for choices when they're missing", () => {
+    const newData = baseAuthoringProps.preprocessFormData({
+      version: 1,
+      prompt: "Test prompt",
+      hint: "Test instructions",
+      choices: [
+        {content: "A", correct: true} as IChoice,
+        {content: "B", correct: true} as IChoice
+      ]
+    });
+    const choices = newData.choices!;
 
-describe("App", () => {
-  beforeEach(() => {
-    mode = undefined;
+    expect(choices[0].id).toBeDefined();
+    expect(choices[1].id).toBeDefined();
+    expect(choices[0].id).not.toEqual(choices[1].id);
   });
 
-  it("should render Runtime or Authoring component depending on the mode", () => {
-    mode = "runtime";
-    let wrapper = shallow(<App />);
-    expect(wrapper.find(Authoring).length).toEqual(0);
-    expect(wrapper.find(Runtime).length).toEqual(1);
-
-    mode = "authoring";
-    wrapper = shallow(<App />);
-    expect(wrapper.find(Authoring).length).toEqual(1);
-    expect(wrapper.find(Runtime).length).toEqual(0);
-
-    mode = "report";
-    wrapper = shallow(<App />);
-    expect(wrapper.find(Authoring).length).toEqual(0);
-    expect(wrapper.find(Runtime).length).toEqual(1);
-    expect(wrapper.find(Runtime).props().report).toEqual(true);
+  it("doesn't overwrite existing choice IDs", () => {
+    expect(baseAuthoringProps.preprocessFormData({
+      version: 1,
+      prompt: "Test prompt",
+      hint: "Test instructions",
+      choices: [
+        {id: "1", content: "A", correct: true},
+        {id: "2", content: "B", correct: true}
+      ]
+    })).toEqual({
+      version: 1,
+      prompt: "Test prompt",
+      hint: "Test instructions",
+      choices: [
+        {id: "1", content: "A", correct: true},
+        {id: "2", content: "B", correct: true}
+      ]
+    });
   });
 });
