+++ src/open-response/components/app.tsx
@@ -1,0 +1,73 @@
+import React from "react";
+import { BaseQuestionApp } from "../../shared/components/base-question-app";
+import { Runtime } from "./runtime";
+import { JSONSchema6 } from "json-schema";
+
+// Note that TS interfaces should match JSON schema. Currently there's no way to generate one from the other.
+// TS interfaces are not available in runtime in contrast to JSON schema.
+
+export interface IAuthoredState {
+  version: number;
+  prompt?: string;
+  defaultAnswer?: string;
+  hint?: string;
+  required?: boolean;
+}
+
+export interface IInteractiveState {
+  response?: string;
+  submitted?: boolean;
+}
+
+const baseAuthoringProps = {
+  schema: {
+    type: "object",
+    properties: {
+      version: {
+        type: "number",
+        default: 1
+      },
+      prompt: {
+        title: "Prompt",
+        type: "string"
+      },
+      required: {
+        title: "Required",
+        type: "boolean"
+      },
+      hint: {
+        title: "Hint",
+        type: "string"
+      },
+      defaultAnswer: {
+        type: "string",
+        title: "Default answer"
+      }
+    }
+  } as JSONSchema6,
+
+  uiSchema: {
+    version: {
+      "ui:widget": "hidden"
+    },
+    prompt: {
+      "ui:widget": "textarea"
+    },
+    hint: {
+      "ui:widget": "textarea"
+    },
+    defaultAnswer: {
+      "ui:widget": "textarea"
+    },
+  }
+};
+
+const isAnswered = (interactiveState: IInteractiveState | undefined) => !!interactiveState?.response;
+
+export const App = () => (
+  <BaseQuestionApp<IAuthoredState, IInteractiveState>
+    Runtime={Runtime}
+    baseAuthoringProps={baseAuthoringProps}
+    isAnswered={isAnswered}
+  />
+);
