+++ src/scaffolded-question/components/app.tsx
@@ -1,0 +1,99 @@
+import React from "react";
+import { JSONSchema6 } from "json-schema";
+import { BaseQuestionApp } from "../../shared/components/base-question-app";
+import { Runtime } from "./runtime";
+import { IframeAuthoring } from "./iframe-authoring";
+
+// Note that TS interfaces should match JSON schema. Currently there's no way to generate one from the other.
+// TS interfaces are not available in runtime in contrast to JSON schema.
+
+export interface IAuthoredState {
+  version: number;
+  prompt?: string;
+  required?: boolean;
+  hint?: string;
+  subinteractives: {
+    id: string;
+    url: string;
+    authoredState: any;
+  }[]
+}
+
+export interface IInteractiveState {
+  subinteractiveStates: {
+    [id: string]: any;
+  },
+  currentSubinteractiveId: string;
+  submitted: boolean;
+}
+
+const baseAuthoringProps = {
+  schema: {
+    type: "object",
+    properties: {
+      version: {
+        type: "number",
+        default: 1
+      },
+      prompt: {
+        title: "Prompt",
+        type: "string"
+      },
+      required: {
+        title: "Required",
+        type: "boolean"
+      },
+      hint: {
+        title: "Hint",
+        type: "string"
+      },
+      subinteractives: {
+        type: "array",
+        title: "Subquestions",
+        items: {
+          type: "object",
+          properties: {
+            id: {
+              type: "string"
+            },
+            url: {
+              type: "string"
+            },
+            authoredState: {
+              type: "any"
+            }
+          }
+        }
+      }
+    }
+  } as JSONSchema6,
+
+  uiSchema: {
+    version: {
+      "ui:widget": "hidden"
+    },
+    prompt: {
+      "ui:widget": "textarea"
+    },
+    hint: {
+      "ui:widget": "textarea"
+    },
+    subinteractives: {
+      items: {
+        "ui:field": "iframeAuthoring"
+      }
+    }
+  },
+
+  fields: {
+    iframeAuthoring: IframeAuthoring
+  }
+};
+
+export const App = () => (
+  <BaseQuestionApp<IAuthoredState, IInteractiveState>
+    Runtime={Runtime}
+    baseAuthoringProps={baseAuthoringProps}
+    disableSubmitBtnRendering={true}
+  />
+);
