+++ src/shared/styles/helpers.scss
@@ -72,10 +72,3 @@ $themeColor2: #34a5be;
   height: 22px;
   vertical-align: middle;
 }
-
-.extraInstructions {
-  margin-top: 10px;
-  background: #f5f5f5;
-  padding: 5px;
-  font-size: 0.9em;
-}
